Restructure default folder
