+++ src/interpreter/default/float/index.ts
@@ -1,0 +1,8 @@
+import { ceil } from './ceil';
+
+export const floatDefaultFunctions = {
+  ceil: {
+    value: ceil,
+    type: ceil.getType(),
+  },
+};
