+++ src/interpreter/default/string/index.ts
@@ -1,0 +1,8 @@
+import { length } from './length';
+
+export const stringDefaultFunctions = {
+  length: {
+    value: length,
+    type: length.getType(),
+  },
+};
