+++ src/interpreter/default/string/length.ts
@@ -3,8 +3,8 @@ import assert from 'assert';
 import { StringWrapper } from 'parser/wrappers';
 import { intType, makeFunctionType, stringType } from 'types/utils';
 
-import { RuntimeResult } from '../../runtimeTypes';
-import { DefaultClosure } from '../closure';
+import { RuntimeResult } from '../../../runtimeTypes';
+import { DefaultClosure } from '../../closure';
 
 export const length = new DefaultClosure(
   makeFunctionType(stringType, intType),
